Updates for 10.5 API changes
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@27962 72102866-910b-0410-8b05-ffd578937521

+++ Headers/Additions/GNUstepBase/GSMime.h
@@ -62,7 +62,7 @@ extern "C" {
 }
 - (BOOL) atEnd;
 - (BOOL) decodeData: (const void*)sData
-             length: (unsigned)length
+             length: (NSUInteger)length
 	   intoData: (NSMutableData*)dData;
 - (void) setAtEnd: (BOOL)flag;
 @end
