+++ Headers/Foundation/NSURLCache.h
@@ -137,32 +137,32 @@ typedef enum
  * Returns the current size (butes) of the data stored in the on-disk
  * cache.
  */
-- (unsigned) currentDiskUsage;
+- (NSUInteger) currentDiskUsage;
 
 /**
  * Returns the current size (butes) of the data stored in the in-memory
  * cache.
  */
-- (unsigned) currentMemoryUsage;
+- (NSUInteger) currentMemoryUsage;
 
 /**
  * Returns the disk capacity (in bytes) of the cache.
  */
-- (unsigned) diskCapacity;
+- (NSUInteger) diskCapacity;
 
 /**
  * Returns the receiver initialised with the specified capacities
  * (in bytes) and using the specified location on disk for persistent
  * storage.
  */
-- (id) initWithMemoryCapacity: (unsigned)memoryCapacity
-		 diskCapacity: (unsigned)diskCapacity
+- (id) initWithMemoryCapacity: (NSUInteger)memoryCapacity
+		 diskCapacity: (NSUInteger)diskCapacity
 		     diskPath: (NSString *)path;
 
 /**
  * Returns the memory capacity (in bytes) of the cache.
  */
-- (unsigned) memoryCapacity;
+- (NSUInteger) memoryCapacity;
 
 /**
  * Empties the cache.
@@ -178,12 +178,12 @@ typedef enum
 /**
  * Sets the disk capacity (in bytes) truncating cache contents if necessary.
  */
-- (void) setDiskCapacity: (unsigned)diskCapacity;
+- (void) setDiskCapacity: (NSUInteger)diskCapacity;
 
 /**
  * Sets the memory capacity (in bytes) truncating cache contents if necessary.
  */
-- (void) setMemoryCapacity: (unsigned)memoryCapacity;
+- (void) setMemoryCapacity: (NSUInteger)memoryCapacity;
 
 /**
  * Stores cachedResponse in the cache, keyed on request.<br />
